Fix OrderedTableDataset and address review comments
