+++ cdap-data-fabric/src/test/java/co/cask/cdap/data2/dataset2/lib/table/inmemory/InMemoryTableTest.java
@@ -19,7 +19,7 @@ package co.cask.cdap.data2.dataset2.lib.table.inmemory;
 import co.cask.cdap.api.dataset.DatasetAdmin;
 import co.cask.cdap.api.dataset.DatasetProperties;
 import co.cask.cdap.api.dataset.table.ConflictDetection;
-import co.cask.cdap.data2.dataset2.lib.table.ordered.BufferingTableTest;
+import co.cask.cdap.data2.dataset2.lib.table.BufferingTableTest;
 
 /**
  *
